Fix recursivity handling.
We were using a global map where we stored all the fields for every struct; this ensured that all structs were processed only once (correctly preventing recursivity loops), but had the unpleasant effect that only _the first_ processed field was injected in _every place_ its type was used.

This fixes that, by handling recursivity with a stack that grows and dies with the call stack, thus preventing that bug. (Plus, is significantly simpler.)

Oh, and tests pass now.